Major refactor: change file structure, change backend to websocket - WIP
