+++ src/assets/IconTemplate@2x.png
