+++ src/assets/IconWinAlt.png
