+++ src/ipc.js
@@ -1,0 +1,70 @@
+import { app, ipcMain } from "electron";
+import Bonjour from "bonjour-service";
+const bonjour = new Bonjour.Bonjour();
+import config from '../config.js';
+import { currentInstance } from './instance.js';
+import { reinitMainWindow } from "./window.js";
+
+ipcMain.on("get-instances", async (event) => {
+  const instances = await getBonjourResult(config.get("allInstances") || []);
+  event.reply("receive-instances", instances);
+});
+
+ipcMain.on("get-ha-instance", (event, url) => {
+  if (url) {
+    addInstance(url);
+  }
+
+  if (currentInstance()) {
+    event.reply("receive-ha-instance", currentInstance());
+  }
+});
+
+ipcMain.on("reconnect", async () => {
+  await reinitMainWindow();
+});
+
+ipcMain.on("restart", () => {
+  app.relaunch();
+  app.exit();
+});
+
+async function getBonjourResult(instances) {
+  return new Promise((resolve) => {
+    const foundInstances = [];
+    
+    bonjour.find({ type: "home-assistant" }, (instance) => {
+      if (instance.txt.internal_url && instances.indexOf(instance.txt.internal_url) === -1) {
+        foundInstances.push(instance.txt.internal_url);
+      }
+      if (instance.txt.external_url && instances.indexOf(instance.txt.external_url) === -1) {
+        foundInstances.push(instance.txt.external_url);
+      }
+    });
+    setTimeout(() => {
+      resolve(foundInstances);
+    }, 1500);
+  });
+}
+
+function addInstance(url) {
+  if (!config.has("allInstances")) {
+    config.set("allInstances", []);
+  }
+
+  const instances = config.get("allInstances");
+
+  if (instances.find((e) => e === url)) {
+    currentInstance(url);
+
+    return;
+  }
+
+  if (!instances.length) {
+    config.set("disableHover", false);
+  }
+
+  instances.push(url);
+  config.set("allInstances", instances);
+  currentInstance(url);
+}
