+++ web/preload.cjs
@@ -12,5 +12,5 @@ contextBridge.exposeInMainWorld('electron', {
     restart: () => ipcRenderer.send('restart'),
     onRetryUpdate: (callback) => ipcRenderer.on('retry-update', (event, data) => callback(data)),
     onRetrySuccess: (callback) => ipcRenderer.on('retry-success', (event, data) => callback(data)),
-    onRetryError: (callback) => ipcRenderer.on('retry-error', (event, data) => callback(data))
+    onRetryError: (callback) => ipcRenderer.on('retry-error', (event, data) => callback(data)),
 });
